pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preen"
version = "0.1.0"
description = "An opinionated, agentic CLI for Python package hygiene and release"
authors = [
    { name = "Gaurav Sood", email = "contact@gsood.com" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9",
    "rich>=13.5",
    "PyYAML>=6.0",
    "Jinja2>=3.0",
    "httpx>=0.24.0",
    "pydoclint>=0.3.0",
    "pyright>=1.1.350",
    "codespell>=2.3.0",
    "ruff>=0.14.0",
]
keywords = ["packaging", "release", "metadata", "hygiene", "cli", "python"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
"Homepage" = "https://github.com/gojiplus/preen"
"Repository" = "https://github.com/gojiplus/preen"
"Documentation" = "https://gojiplus.github.io/preen"
"Bug Tracker" = "https://github.com/gojiplus/preen/issues"
"Changelog" = "https://github.com/gojiplus/preen/releases"

[project.scripts]
preen = "preen.cli:run"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
    "furo>=2024.0",
]

[tool.setuptools.packages.find]
include = ["preen", "preen.*"]

[tool.pytest.ini_options]
addopts = "-q"
